+++ src/components/common/Auth/AuthInputNormal/index.tsx
@@ -5,7 +5,7 @@ import { AuthInputNormal, AuthInputStyleProps } from '@/types/auth';
 import closeIcon from '@assets/icons/ic-login-close.svg';
 import checkInvalid from '@assets/icons/ic-signup-check-invalid.svg';
 import checkValid from '@assets/icons/ic-signup-check-valid.svg';
-import { fieldOptions } from '@utils/lib/login';
+import { fieldOptions } from '@utils/lib/auth';
 
 const AuthInputNormal = ({
   type,
